+++ world/Data.h
@@ -120,7 +120,7 @@ namespace Data {
 		String mSuffix = "";
 		bool mGM = false;
 		bool mMuted = false;
-		u8 mStatus = 0;
+		u32 mStatus = 0;
 
 		u16 mZoneID = 0; // TODO: Choose a good default.
 		u16 mInstanceID = 0;
